// src/redux/types/userTypes.tsx
export enum USER_TYPES{
    GET_USERS_FETCH = "GET_USERS_FETCH",
    GET_USERS_LOADING = "GET_USERS_LOADING",
    GET_USERS_SUCCESS = "GET_USERS_SUCCESS",
    GET_USERS_FAILED = "GET_USERS_FAILED",
    CREATE_USER_FETCH = "CREATE_USER_FETCH",
    CREATE_USER_LOADING = "CREATE_USER_LOADING",
    CREATE_USER_SUCCESS = "CREATE_USER_SUCCESS",
    CREATE_USER_FAILED = "CREATE_USER_FAILED",
    SIGN_OUT_FETCH = "SIGN_OUT_FETCH",
    SIGN_OUT_LOADING = "SIGN_OUT_LOADING",
    SIGN_OUT_SUCCESS = "SIGN_OUT_SUCCESS",
    SIGN_OUT_FAILED = "SIGN_OUT_FAILED"
}